+++ nacc_attribute_deriver/attributes/derived/modules/__init__.py
@@ -1,4 +1,5 @@
 # ruff: noqa
+from .b1a import *
 from .cls import *
 from .csf import *
 from .ftld import *